pr fixes
split the css of the button and the window into 2 files
