+++ src/Screens/Gameplay/Components/Seen/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { GeneralButton } from 'GeneralComponents/GeneralButton/styles';
-import { discoveryButtonSize } from 'Screens/Gameplay/Components/Discovery/styles';
+import { discoveryButtonSize } from 'Screens/Gameplay/Components/Discovery/styles_button';
 import { animated } from 'react-spring';
 
 export const seenButtonSize = 1.5;
